fix annotation reactivity
