+++ .changeset/afraid-countries-sort.md
@@ -1,0 +1,5 @@
+---
+'@evidence-dev/core-components': patch
+---
+
+Fix ReferenceLine and ReferenceArea reactivity
